l1: convert happy flow test into integration test
MOVE only, no logic changes except making `send_events_to_l1_provider`
pub.
